Try to clear up some of the remaining sporadic failures (#2394)
* Try to clear up some of the remaining sporadic failures

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* check about deleting file first

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fix some warnings and try again on logging_file tests

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

+++ tests/helics/shared_library/loggingTests.cpp
@@ -205,51 +205,58 @@ TEST(logging_tests, broker_logging)
 
 TEST(logging_tests, broker_logging_file)
 {
-    const char* lfile = "logb.txt";
-    std::filesystem::remove(lfile);
+    const std::string lfile{"logb.txt"};
+
+    if (std::filesystem::exists(lfile)) {
+        std::filesystem::remove(lfile);
+    }
     auto err = helicsErrorInitialize();
     auto broker = helicsCreateBroker("inproc", "blog", "--log_level=trace", &err);
-    helicsBrokerSetLogFile(broker, lfile, &err);
+    helicsBrokerSetLogFile(broker, lfile.c_str(), &err);
     helicsBrokerDisconnect(broker, &err);
     helicsCloseLibrary();
-    EXPECT_TRUE(std::filesystem::exists(lfile));
+    ASSERT_TRUE(std::filesystem::exists(lfile));
     std::filesystem::remove(lfile);
     EXPECT_EQ(err.error_code, 0);
 }
 
 TEST(logging_tests, core_logging_file)
 {
-    const char* lfile = "logc.txt";
-    std::filesystem::remove(lfile);
+    const std::string lfile{"logc.txt"};
+    if (std::filesystem::exists(lfile)) {
+        std::filesystem::remove(lfile);
+    }
     auto core = helicsCreateCore("inproc", "clog", "--autobroker --log_level=trace", nullptr);
 
     auto err = helicsErrorInitialize();
-    helicsCoreSetLogFile(core, lfile, &err);
+    helicsCoreSetLogFile(core, lfile.c_str(), &err);
     helicsCoreDisconnect(core, &err);
     helicsCloseLibrary();
-    EXPECT_TRUE(std::filesystem::exists(lfile));
+    ASSERT_TRUE(std::filesystem::exists(lfile));
     std::filesystem::remove(lfile);
 }
 
 TEST(logging_tests, fed_logging_file)
 {
-    const char* lfile = "logf.txt";
-    std::filesystem::remove(lfile);
+    const std::string lfile{"logf.txt"};
+    if (std::filesystem::exists(lfile)) {
+        std::filesystem::remove(lfile);
+    }
     auto core = helicsCreateCore("inproc", "clogf", "--autobroker --log_level=trace", nullptr);
 
     auto err = helicsErrorInitialize();
     auto fi = helicsCreateFederateInfo();
     helicsFederateInfoSetCoreName(fi, "clogf", nullptr);
     auto fed = helicsCreateValueFederate("f1", fi, nullptr);
-    helicsFederateSetLogFile(fed, lfile, nullptr);
+    helicsFederateSetLogFile(fed, lfile.c_str(), nullptr);
 
-    helicsCoreSetLogFile(core, lfile, &err);
+    helicsCoreSetLogFile(core, lfile.c_str(), &err);
     helicsCoreDisconnect(core, &err);
     helicsFederateFinalize(fed, &err);
 
     helicsFederateSetLogFile(fed, "emptyfile.txt", nullptr);
     helicsFederateInfoFree(fi);
     helicsCloseLibrary();
-    EXPECT_TRUE(std::filesystem::exists(lfile));
+    ASSERT_TRUE(std::filesystem::exists(lfile));
     std::filesystem::remove(lfile);
 }
